добавил поля в бд вынес в отдельный файл работу с sql
