+++ main.py
@@ -1,41 +1,8 @@
 import telebot
 from telebot import types
-import sqlite3
+from sql.use_sql import *
 
 bot = telebot.TeleBot('5162531568:AAFulbpqupsSMHiri53UD0jIRC7gpzUayTc')
-
-
-def sql_new_user(tg_id, tg_username):
-    """
-    Функция реализует создание записи в базе данных о пользователе
-    :param tg_id:
-    :param tg_username:
-    :return: nothing
-    """
-    conn = sqlite3.connect('EnglishBotka.db')
-    cursor = conn.cursor()
-    req = f"INSERT INTO User VALUES ({tg_id}, '{tg_username}', 0)"
-    cursor.execute(req)
-    conn.close()
-
-
-def sql_is_user_in_db(tg_id):
-    """
-    Функциф проверяет создан существует ли уже такой юзер в базе данных
-    :param tg_id:
-    :return: true/false
-    """
-    conn = sqlite3.connect('EnglishBotka.db')
-    cursor = conn.cursor()
-    req = f"SELECT tg_id FROM User"
-    cursor.execute(req)
-    result = cursor.fetchall()
-    conn.close()
-    for i in result:
-        if str(tg_id) in i:
-            return True
-    return False
-
 
 
 @bot.message_handler(commands=['start'])
